created tenpercentdiscount class implements discount interface

+++ app/src/main/java/com/example/huascar/shoppingbasket/Interfaces/Discount.java
@@ -6,5 +6,5 @@ package com.example.huascar.shoppingbasket.Interfaces;
 
 public interface Discount {
 
-    double Discount(int total);
+    double Discount(double total);
 }
